webapp/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
boto3==1.42.89
botocore==1.42.89
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==1.3.3
contourpy==1.3.3
cssselect==1.4.0
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
deepface==0.0.99
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.135.3
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.28.0
fire==0.7.1
Flask==2.2.5
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
fuzzywuzzy==0.18.0
gast==0.7.0
gdown==6.0.0
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
imutils==0.5.4
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.14.0
kiwisolver==1.5.0
lazy-loader==0.5
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.23
locationtagger==0.0.1
lxml==6.0.4
lxml_html_clean==0.4.4
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
murmurhash==1.0.15
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
newspaper3k==0.2.8
ninja==1.13.0
nltk==3.8.1
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
parso==0.8.6
pdf2image==1.17.0
pillow==12.2.0
platformdirs==4.9.6
polars==1.39.3
polars-runtime-32==1.39.3
preshed==3.0.13
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyclipper==1.4.0
pycountry==26.2.16
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pyodbc==5.3.0
pyparsing==3.3.2
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
retina-face==0.0.17
rich==15.0.0
s3transfer==0.16.0
scikit-image==0.26.0
scipy==1.17.1
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
thinc==8.3.13
tifffile==2026.3.3
tinysegmenter==0.3
tldextract==5.3.1
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.44.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
Werkzeug==3.1.8
wrapt==2.1.2
